Refactor main.py into modular router structure
Split monolithic main.py into separate router modules organized by domain:

- routes/__init__.py - Package initialization and router exports
- routes/auth.py - Authentication and login routes (login, logout, player-name validation)
- routes/lobby.py - Multiplayer lobby routes (lobby page, select/accept/decline opponents, long-polling)
- routes/ship_placement.py - Ship placement routes (place/remove/random ships, ready state)
- routes/gameplay.py - Active game gameplay routes (game page)
- routes/start_game.py - Game start confirmation routes
- routes/types.py - Type hints for routes package

main.py now serves as composition root, importing and including all routers
while keeping testing endpoints (/test/*) and health check in main app.

Key architecture changes:
- Each router uses module-level service references set during initialization
- Shared services (game_service, lobby_service, templates) passed via set_up_* functions
- All routes organized by domain for better maintainability
- Reduced main.py from 1400 lines to ~150 lines

Tests verified:
- 210 unit tests pass
- 119 endpoint tests pass
- BDD tests pass
