+++ routes/types.py
@@ -1,0 +1,10 @@
+"""Type hints for routes package.
+
+These are forward references used to avoid circular imports
+between routes and the game/services modules.
+"""
+
+from typing import TYPE_CHECKING
+
+if TYPE_CHECKING:
+    pass  # No forward references needed - imports are done directly
